+++ contracts/interfaces/IGovernance.sol
@@ -5,5 +5,5 @@ pragma solidity 0.7.6;
 interface IGovernance {
   function contractGuards(address target) external view returns (address guard);
 
-  function assetGuards(uint8 assetType) external view returns (address guard);
+  function assetGuards(uint16 assetType) external view returns (address guard);
 }
